requirements.txt
absl-py==2.1.0
aiohappyeyeballs==2.4.4
aiohttp==3.11.10
aioitertools==0.12.0
aiosignal==1.3.1
annotated-types==0.7.0
antlr4-python3-runtime==4.9.3
appnope @ file:///home/conda/feedstock_root/build_artifacts/appnope_1733332318622/work
ase==3.23.0
asttokens @ file:///home/conda/feedstock_root/build_artifacts/asttokens_1733250440834/work
astunparse==1.6.3
attrs==24.2.0
azure-core==1.32.0
azure-identity==1.19.0
azure-storage-blob==12.24.0
bcrypt==4.2.1
boto3==1.35.77
botocore==1.35.77
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
chgnet==0.4.0
click==8.1.7
cloudpickle==3.1.0
comm @ file:///home/conda/feedstock_root/build_artifacts/comm_1710320294760/work
ConfigArgParse==1.7
contourpy==1.3.1
cryptography==44.0.0
cycler==0.12.1
Cython==3.0.11
debugpy @ file:///Users/runner/miniforge3/conda-bld/debugpy_1732236743159/work
decorator @ file:///home/conda/feedstock_root/build_artifacts/decorator_1733236420667/work
Deprecated==1.2.15
dnspython==2.7.0
docker-pycreds==0.4.0
e3nn==0.5.0
emmet-core==0.83.9
exceptiongroup @ file:///home/conda/feedstock_root/build_artifacts/exceptiongroup_1733208806608/work
executing @ file:///home/conda/feedstock_root/build_artifacts/executing_1725214404607/work
fairchem-core==1.3.0
filelock==3.16.1
flatbuffers==24.3.25
fonttools==4.55.2
fortls==3.1.2
frozenlist==1.5.0
fsspec==2024.10.0
gast==0.6.0
gitdb==4.0.11
GitPython==3.1.43
google-pasta==0.2.0
grpcio==1.68.1
h5py==3.12.1
hydra-core==1.3.2
idna==3.10
importlib_metadata @ file:///home/conda/feedstock_root/build_artifacts/importlib-metadata_1733223117029/work
iniconfig==2.0.0
ipykernel @ file:///Users/runner/miniforge3/conda-bld/ipykernel_1719845458456/work
ipython @ file:///home/conda/feedstock_root/build_artifacts/ipython_1732896932739/work
isodate==0.7.2
jedi @ file:///home/conda/feedstock_root/build_artifacts/jedi_1733300866624/work
Jinja2==3.1.4
jmespath==1.0.1
joblib==1.4.2
json5==0.10.0
jsonlines==4.0.0
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
jupyter_client @ file:///home/conda/feedstock_root/build_artifacts/jupyter_client_1733440914442/work
jupyter_core @ file:///home/conda/feedstock_root/build_artifacts/jupyter_core_1727163409502/work
keras==3.7.0
kiwisolver==1.4.7
latexcodec==3.0.0
libclang==18.1.1
lightning-utilities==0.11.9
llvmlite==0.43.0
lmdb==1.5.1
loguru==0.7.3
m3gnet==0.2.4
mace-torch==0.3.8
maggma==0.70.0
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.9.3
matplotlib-inline @ file:///home/conda/feedstock_root/build_artifacts/matplotlib-inline_1733416936468/work
matscipy==1.1.1
mattersim==1.0.0rc9
mdurl==0.1.2
ml-dtypes==0.3.2
mongomock==4.3.0
monty==2024.10.21
mp-api==0.43.0
mpmath==1.3.0
msal==1.31.1
msal-extensions==1.2.0
msgpack==1.1.0
multidict==6.1.0
namex==0.0.8
nest_asyncio @ file:///home/conda/feedstock_root/build_artifacts/nest-asyncio_1733325553580/work
networkx==3.4.2
numba==0.60.0
numpy==1.26.4
nvidia-ml-py3==7.352.0
omegaconf==2.3.0
opt-einsum-fx==0.1.4
opt_einsum==3.4.0
optree==0.13.1
orjson==3.10.12
packaging @ file:///home/conda/feedstock_root/build_artifacts/packaging_1733203243479/work
palettable==3.3.3
pandas==2.2.3
paramiko==3.5.0
parso @ file:///home/conda/feedstock_root/build_artifacts/parso_1733271261340/work
pexpect @ file:///home/conda/feedstock_root/build_artifacts/pexpect_1733301927746/work
pickleshare @ file:///home/conda/feedstock_root/build_artifacts/pickleshare_1733327343728/work
pillow==11.0.0
platformdirs @ file:///home/conda/feedstock_root/build_artifacts/platformdirs_1733232627818/work
plotly==5.24.1
pluggy==1.5.0
portalocker==2.10.1
prettytable==3.12.0
prompt_toolkit @ file:///home/conda/feedstock_root/build_artifacts/prompt-toolkit_1733302527033/work
propcache==0.2.1
protobuf==4.25.5
psutil @ file:///Users/runner/miniforge3/conda-bld/psutil_1729847073461/work
ptyprocess @ file:///home/conda/feedstock_root/build_artifacts/ptyprocess_1733302279685/work/dist/ptyprocess-0.7.0-py2.py3-none-any.whl#sha256=92c32ff62b5fd8cf325bec5ab90d7be3d2a8ca8c8a3813ff487a8d2002630d1f
pure_eval @ file:///home/conda/feedstock_root/build_artifacts/pure_eval_1721585709575/work
pybtex==0.24.0
pycparser==2.22
pydantic==2.9.2
pydantic-settings==2.6.1
pydantic_core==2.23.4
pydash==8.0.4
Pygments @ file:///home/conda/feedstock_root/build_artifacts/pygments_1733221634316/work
PyJWT==2.10.1
pymatgen==2024.4.13
pymongo==4.10.1
PyNaCl==1.5.0
pyparsing==3.2.0
pytest==8.3.4
python-dateutil @ file:///home/conda/feedstock_root/build_artifacts/python-dateutil_1733215673016/work
python-dotenv==1.0.1
python-hostlist==2.2.1
pytz==2024.2
PyYAML==6.0.2
pyzmq @ file:///Users/runner/miniforge3/conda-bld/pyzmq_1728642247675/work
referencing==0.35.1
requests==2.32.3
rich==13.9.4
rpds-py==0.22.3
ruamel.yaml==0.18.6
ruamel.yaml.clib==0.2.12
s3transfer==0.10.4
scikit-learn==1.6.0
scipy==1.14.1
seekpath==2.1.0
sentinels==1.0.0
sentry-sdk==2.19.1
setproctitle==1.3.4
setuptools==75.6.0
six @ file:///home/conda/feedstock_root/build_artifacts/six_1733380938961/work
smart-open==7.0.5
smmap==5.0.1
spglib==2.5.0
sshtunnel==0.4.0
stack-data @ file:///home/conda/feedstock_root/build_artifacts/stack_data_1669632077133/work
submitit==1.5.2
sympy==1.13.1
tabulate==0.9.0
tenacity==9.0.0
tensorboard==2.16.2
tensorboard-data-server==0.7.2
tensorflow==2.16.2
tensorflow-macos==2.16.2
termcolor==2.5.0
threadpoolctl==3.5.0
torch==2.5.1
torch-ema==0.3
torch-runstats==0.2.0
torch_geometric==2.5.3
torch_scatter==2.1.2
torch_sparse==0.6.18
torchaudio==2.2.0
torchmetrics==1.6.0
torchvision==0.17.0
tornado @ file:///Users/runner/miniforge3/conda-bld/tornado_1732615911982/work
tqdm==4.67.1
traitlets @ file:///home/conda/feedstock_root/build_artifacts/traitlets_1733367359838/work
typing_extensions @ file:///home/conda/feedstock_root/build_artifacts/typing_extensions_1733188668063/work
tzdata==2024.2
uncertainties==3.2.2
urllib3==2.2.3
wandb==0.19.0
wcwidth @ file:///home/conda/feedstock_root/build_artifacts/wcwidth_1733231326287/work
Werkzeug==3.1.3
wheel==0.45.1
wrapt==1.17.0
yarl==1.18.3
zipp @ file:///home/conda/feedstock_root/build_artifacts/zipp_1732827521216/work
